CipherCode_Lexora/rag-backend/requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
unstructured
python-dotenv
tqdm
loguru
sentence-transformers
langchain-groq
fastapi
uvicorn
